Revert "v.0.0.8] feat. CharacterDeath 추가 중"
This reverts commit 37240ee7148c6a3281424267b0e2200e9923acf3.

+++ Assets/3.Script/Characters/Players/PlayerManagers/PlayerUIHudManager.cs
@@ -14,7 +14,7 @@ namespace KF
             healthBar.gameObject.SetActive(true);
         }
 
-        public void SetNewHeathValue(int oldValue, int newValue)
+        public void SetNewHeathValue(float oldValue, float newValue)
         {
             healthBar.SetStat(Mathf.RoundToInt(newValue));
         }
